Move smispy plugin
Move the smispy local plugin files to $(pythondir)/smispy_plugin
ref. https://github.com/libstorage/libstoragemgmt/issues/440
