+++ plugin/smispy_plugin/__init__.py
@@ -1,0 +1,1 @@
+from smispy_plugin.smis import Smis
